rename idf image format, don't rexport things we don't need to
